Updated dependencies. Reorganized some files. Modernized the ESLint setup.
