--- .eslintrc
@@ -1,11 +1,0 @@
-{
-	"extends" : "protect-me-from-my-stupidity",
-
-	"env" :
-	{
-		"browser" : true,
-		"es6" : true
-	},
-
-	"root" : true
-}
