+++ .github/workflows/build-and-deploy.yml
@@ -41,7 +41,7 @@ jobs:
         run: npm run build -- --env baseUri=/cv/
 
       - name: Package
-        uses: actions/upload-pages-artifact@v2
+        uses: actions/upload-pages-artifact@3
         with:
           path: site/
 
@@ -69,4 +69,4 @@ jobs:
 
       - name: Deploy
         id: deployment
-        uses: actions/deploy-pages@v2
+        uses: actions/deploy-pages@v4
