+++ eslint.config.js
@@ -1,0 +1,20 @@
+import protectMeFromMyStupidity             from 'eslint-config-protect-me-from-my-stupidity';
+import andFromWritingStupidNodeApplications from 'eslint-config-protect-me-from-my-stupidity/and/from-writing-stupid-node-applications';
+import andFromWritingStupidWebApplications  from 'eslint-config-protect-me-from-my-stupidity/and/from-writing-stupid-web-applications';
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+export default [
+	{
+		ignores : ['site/**/*']
+	},
+	...protectMeFromMyStupidity(),
+	...andFromWritingStupidWebApplications([
+		'src/scripts/**/*.js'
+	]),
+	...andFromWritingStupidNodeApplications([
+		'*.config.js',
+		'tools/**/*.js',
+		'src/templates/helpers/**/*.js'
+	])
+];
