+++ src/scripts/email-revealer.js
