+++ src/scripts/lazy-image-detector.js
