--- src/templates/.eslintrc
@@ -1,12 +1,0 @@
-{
-	"extends" :
-	[
-		"protect-me-from-my-stupidity/and/from-writing-stupid-node-applications"
-	],
-
-	"env" :
-	{
-		"browser" : false,
-		"node" : true
-	}
-}
